build: warn on the code we write, silence the code we generate
The target carried a blanket `-w -Wno-implicit-function-declaration`, which
suppressed every diagnostic in the project. That is how two calls with no
prototype in scope survived -- C gives those `int f()`, the wrong return type
-- and it means no warning has been actionable here for as long as the flag
has been set.

Measured before changing anything. `-Wall -Wextra` over the 125 hand-written
sources reports 311 warnings, but 301 are two benign categories: 224
missing-field-initializers (219 of them the settings descriptor registry, where
designated initializers are the house style) and 77 unknown-pragma from the
snesrecomp runtime's types.h. The real signal is ten, all fixed here:

  four deprecated ucontext calls -- the coroutine model depends on them and
    macOS offers no replacement, so the deprecation is acknowledged once at
    the top of actraiser_rtl.c instead of at four call sites
  two `/*` inside block comments, one ours and one in the runtime header
  two unused parameters, each with a single caller, so both are removed
    rather than voided
  one SDL_Scancode-versus-int sign comparison in the menu binding lookup

src/gen would add roughly 8,000 more, so the polarity matters: `-w` beats
`-Wall` in clang regardless of order, which means the real warnings go on the
target and `-w` is re-applied per file to the generated banks and the vendored
runner. -Werror=implicit-function-declaration and -Werror=implicit-int are on,
so the class of defect found in the previous commit cannot recur silently.

A full clean build now produces zero compiler warnings, which makes any future
one a defect rather than background noise. code-style.md records that.

Behaviour unchanged: byte-identical Aitos composite in both terrain
configurations, 69/69 tests in each, D7-voxel-town clean.

+++ src/settings_overlay.c
@@ -2436,7 +2436,8 @@ bool SettingsOverlay_HandleGamepadEvent(const SDL_Event *event) {
 static bool MenuKeyMatchesBinding(SDL_Keycode key, InputAction action) {
   uint32 binding = g_settings.input_bind[kInputClass_Keyboard][action];
   if (INPUT_BIND_KIND(binding) != kInputBind_Key) return false;
-  return SDL_GetScancodeFromKey(key, NULL) == INPUT_BIND_CODE(binding);
+  return (int)SDL_GetScancodeFromKey(key, NULL) ==
+      (int)INPUT_BIND_CODE(binding);
 }
 
 /* Maps a keycode to a menu command through the player's OWN keyboard bindings,
